feat: Move MySQL specific types to mysqlPlugin folder (#15784)
