+++ app/server/appsmith-plugins/mysqlPlugin/src/main/java/com/external/plugins/datatypes/MySQLBooleanType.java
@@ -1,4 +1,4 @@
-package com.appsmith.external.datatypes.mysql;
+package com.external.plugins.datatypes;
 
 import com.appsmith.external.constants.DataType;
 import com.appsmith.external.datatypes.AppsmithType;
